refactor(plugins): extract runIsolatedCompletion into @sero-ai/extension-runtime
The memory and web plugins each carried a near-identical copy of the
isolated-completion helper (~65 lines): the delicate part of the Pi 0.80.6
migration (correct provider dispatch, session disposal, cancellation). Two
copies invite silent drift.

- New workspace package @sero-ai/extension-runtime holds the single canonical
  runIsolatedCompletion with a unified flat signature. Pi SDK stays a peer;
  the plugin builder inlines the package and keeps the peers external.
- Both plugins consume it as a workspace:* devDependency and drop their local
  copies. The custom-provider dispatch test moves to the package as its
  canonical home.
- Add the @sero-ai/extension-runtime path mapping to the shared extension
  tsconfig and document the package in AGENTS.md.

Verified: pnpm typecheck (19/19), package + memory + web test suites pass, and
both plugin extension bundles inline the helper with pi-* left external.
