+++ packages/extension-runtime/src/index.ts
@@ -1,0 +1,2 @@
+export { runIsolatedCompletion } from './isolated-completion';
+export type { IsolatedCompletionRequest } from './isolated-completion';
