+++ plugins/sero-web-plugin/extension/summary-review.ts
@@ -1,6 +1,6 @@
 import type { Api, Model } from "@earendil-works/pi-ai";
 import type { ExtensionContext } from "@earendil-works/pi-coding-agent";
-import { runIsolatedCompletion } from "./isolated-completion.js";
+import { runIsolatedCompletion } from "@sero-ai/extension-runtime";
 import type { QueryResultData } from "./storage.js";
 
 const PREFERRED_SUMMARY_MODELS = [
