+++ plugins/sero-web-plugin/extension/tsconfig.json
@@ -12,7 +12,6 @@
     "./gemini-search.ts",
     "./gemini-search-config.ts",
     "./gemini-search-format.ts",
-    "./isolated-completion.ts",
     "./summary-review.ts",
     "./video-extract.ts",
     "./video-config.ts",
